pyproject.toml
[project]
name = "textra-mcp-server"
version = "0.1.0"
description = "MCP server for Textra translation API (Ja->En)"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0", # Specify a suitable version constraint
    "requests",
    "requests-oauthlib",
    "python-dotenv",
]
# Add other metadata like authors, license, etc. as needed

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "ruff",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function" # Add this line

# Use [tool.scripts] for uv run compatibility
[tool.scripts]
test = "pytest"

[tool.setuptools]
py-modules = ["config", "server", "textra_client"]
